refactor: 피드백 포함 수정 & 폴더 분리
